# Optimize srgb decoding