--- a/srgb/srgb_test.go
+++ b/srgb/srgb_test.go
@@ -24,7 +24,28 @@ func TestDecode(t *testing.T) {
 	}
 }
 
-func BenchmarkDecode(b *testing.B) {
+func BenchmarkDecode_RGBA(b *testing.B) {
+	input := image.NewRGBA(image.Rect(0, 0, 512, 512))
+	for i := 0; i < b.N; i++ {
+		Decode(input)
+	}
+}
+
+func BenchmarkDecode_RGBA64(b *testing.B) {
+	input := image.NewRGBA64(image.Rect(0, 0, 512, 512))
+	for i := 0; i < b.N; i++ {
+		Decode(input)
+	}
+}
+
+func BenchmarkDecode_NRGBA(b *testing.B) {
+	input := image.NewNRGBA(image.Rect(0, 0, 512, 512))
+	for i := 0; i < b.N; i++ {
+		Decode(input)
+	}
+}
+
+func BenchmarkDecode_NRGBA64(b *testing.B) {
 	input := image.NewNRGBA64(image.Rect(0, 0, 512, 512))
 	for i := 0; i < b.N; i++ {
 		Decode(input)