```
goos: darwin
goarch: arm64
pkg: github.com/shogo82148/go-imaging/srgb
                  │   .old.txt   │              .new.txt               │
                  │    sec/op    │   sec/op     vs base                │
Decode_RGBA-10      2169.4µ ± 1%   489.2µ ± 1%  -77.45% (p=0.000 n=10)
Decode_RGBA64-10    1622.2µ ± 7%   448.6µ ± 1%  -72.35% (p=0.000 n=10)
Decode_NRGBA-10     2174.1µ ± 3%   486.1µ ± 1%  -77.64% (p=0.000 n=10)
Decode_NRGBA64-10   1297.5µ ± 1%   457.6µ ± 1%  -64.73% (p=0.000 n=10)
geomean              1.775m        470.0µ       -73.52%

                  │   .old.txt   │               .new.txt               │
                  │     B/op     │     B/op      vs base                │
Decode_RGBA-10      6.003Mi ± 0%   2.001Mi ± 0%  -66.66% (p=0.000 n=10)
Decode_RGBA64-10    6.004Mi ± 0%   2.001Mi ± 0%  -66.66% (p=0.000 n=10)
Decode_NRGBA-10     6.003Mi ± 0%   2.001Mi ± 0%  -66.66% (p=0.000 n=10)
Decode_NRGBA64-10   4.003Mi ± 0%   2.001Mi ± 0%  -50.00% (p=0.000 n=10)
geomean             5.425Mi        2.001Mi       -63.11%

                  │    .old.txt    │              .new.txt               │
                  │   allocs/op    │ allocs/op   vs base                 │
Decode_RGBA-10      524312.00 ± 0%   24.00 ± 0%  -100.00% (p=0.000 n=10)
Decode_RGBA64-10    524314.00 ± 0%   24.00 ± 0%  -100.00% (p=0.000 n=10)
Decode_NRGBA-10     524312.00 ± 0%   24.00 ± 0%  -100.00% (p=0.000 n=10)
Decode_NRGBA64-10   262168.00 ± 0%   24.00 ± 0%   -99.99% (p=0.000 n=10)
geomean                440.9k        24.00        -99.99%
```